postgres impl fix

+++ go-fkc/MemSubmodularCover/utils.go
@@ -228,7 +228,7 @@ func setToBitSet(set map[int]bool, n int) *bitset.BitSet {
 	return bitset
 }
 
-func listToBitSet(list []int, n int) *bitset.BitSet {
+func listToBitSet(list []uint8, n int) *bitset.BitSet {
 	bitset := bitset.New(uint(n))
 	for i := 0; i < len(list); i++ {
 		bitset.Set(uint(list[i]))
